% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_interaction_coefficients}
\alias{extract_interaction_coefficients}
\title{Extract correlational selection coefficients}
\usage{
extract_interaction_coefficients(trait_cols, results)
}
\arguments{
\item{trait_cols}{A character vector of trait column names.}

\item{results}{A model results object returned by \code{analyze_nonlinear_selection()}.}
}
\value{
A data frame with correlational (interaction) selection coefficients and statistics.
}
\description{
Extract correlational selection coefficients
}
