% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_quadratic_coefficients}
\alias{extract_quadratic_coefficients}
\title{Extract quadratic selection coefficients}
\usage{
extract_quadratic_coefficients(trait_cols, results)
}
\arguments{
\item{trait_cols}{A character vector of trait column names.}

\item{results}{A model results object returned by \code{analyze_nonlinear_selection()}.}
}
\value{
A data frame with quadratic selection coefficients (doubled estimates) and statistics.
}
\description{
Extract quadratic selection coefficients
}
