% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adaptive_landscape.R
\name{plot_adaptive_landscape_3d}
\alias{plot_adaptive_landscape_3d}
\title{Plot Adaptive Landscape (3D Perspective)}
\usage{
plot_adaptive_landscape_3d(
  landscape,
  trait_cols,
  theta = -30,
  phi = 30,
  grid_n = 200,
  color_palette = NULL,
  ...
)
}
\arguments{
\item{landscape}{Output object of class \code{"adaptive_landscape"}.}

\item{trait_cols}{Character vector of length 2 specifying the trait column names.}

\item{theta}{Numeric azimuthal viewing angle. Default is -30.}

\item{phi}{Numeric colatitude viewing angle. Default is 30.}

\item{grid_n}{Integer specifying the resolution of the interpolation grid. Default is 200.}

\item{color_palette}{Optional vector of colors for the surface. Defaults to viridis plasma.}

\item{...}{Additional arguments passed to \code{fields::drape.plot()}.}
}
\value{
A 3D plot produced by \code{fields::drape.plot()}.
}
\description{
Plot Adaptive Landscape (3D Perspective)
}
