% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_family.R
\name{detect_family}
\alias{detect_family}
\title{Detect GLM Family from Fitness Column}
\usage{
detect_family(fitness_vector)
}
\arguments{
\item{fitness_vector}{Numeric vector of fitness values.}
}
\value{
An object of class \code{family}.
}
\description{
If fitness takes values {0,1}, returns \code{binomial(link="logit")};
otherwise returns \code{gaussian()}.
}
