% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_linear_coefficients}
\alias{extract_linear_coefficients}
\title{Extract Linear (Directional) Coefficients}
\usage{
extract_linear_coefficients(trait_cols, results)
}
\arguments{
\item{trait_cols}{Character vector of trait names.}

\item{results}{A list from \code{analyze_linear_selection()}.}
}
\value{
A tidy \code{data.frame}.
}
\description{
Extract Linear (Directional) Coefficients
}
